@@ -21,28 +21,42 @@
 
 package jolie.embedding.jolie;
 
-import jolie.cli.CommandLineException;
-import jolie.cli.CommandLineParser;
-import jolie.Interpreter;
-import jolie.runtime.Value;
-import jolie.runtime.embedding.EmbeddedServiceLoader;
-import jolie.runtime.embedding.EmbeddedServiceLoadingException;
-import jolie.runtime.expression.Expression;
-
 import java.io.ByteArrayInputStream;
 import java.io.File;
+import java.io.FileNotFoundException;
 import java.io.IOException;
 import java.util.Optional;
 import java.util.concurrent.ExecutionException;
 import java.util.concurrent.Future;
 import java.util.concurrent.atomic.AtomicLong;
 import java.util.regex.Pattern;
+import jolie.Interpreter;
+import jolie.cli.CommandLineException;
+import jolie.cli.CommandLineParser;
+import jolie.lang.parse.ast.ServiceNode;
+import jolie.lang.parse.module.ModuleSource;
+import jolie.runtime.Value;
+import jolie.runtime.embedding.EmbeddedServiceLoader;
+import jolie.runtime.embedding.EmbeddedServiceLoadingException;
+import jolie.runtime.expression.Expression;
 
 public class JolieServiceLoader extends EmbeddedServiceLoader {
 	private final static Pattern SERVICE_PATH_SPLIT_PATTERN = Pattern.compile( " " );
 	private final static AtomicLong SERVICE_LOADER_COUNTER = new AtomicLong();
 	private final Interpreter interpreter;
 
+	/**
+	 * Creates a new instance of ServiceLoader by file path.
+	 *
+	 * @param channelDest the channel destination to use to communicate with the service loaded by this
+	 *        loader.
+	 * @param currInterpreter the Jolie interpreter to use to load the service.
+	 * @param servicePath the path to the service to load. This path can be a Jolie service file or a
+	 *        directory containing a Jolie service file. If it is a directory, the service file must be
+	 *        named "service.ol".
+	 * @throws IOException if an IO error occurs while loading the service.
+	 * @throws CommandLineException if a command line error occurs while loading the service.
+	 */
 	public JolieServiceLoader( Expression channelDest, Interpreter currInterpreter, String servicePath,
 		Optional< String > serviceName, Optional< Value > params )
 		throws IOException, CommandLineException {
@@ -55,27 +69,50 @@ public class JolieServiceLoader extends EmbeddedServiceLoader {
 
 		System.arraycopy( options, 0, newArgs, 2, options.length );
 		System.arraycopy( ss, 0, newArgs, 2 + options.length, ss.length );
-		CommandLineParser commandLineParser = new CommandLineParser( newArgs, currInterpreter.getClassLoader(), false );
+		try( CommandLineParser commandLineParser = new CommandLineParser( newArgs,
+			currInterpreter.getClassLoader(), false ) ) {
+			Interpreter.Configuration cmdConfig = commandLineParser.getInterpreterConfiguration();
+			Interpreter.Configuration config = Interpreter.Configuration
+				.create(
+					cmdConfig,
+					serviceName.orElse( cmdConfig.executionTarget() ) );
 
-		Interpreter.Configuration cmdConfig = commandLineParser.getInterpreterConfiguration();
-		Interpreter.Configuration config = Interpreter.Configuration.create(
-			cmdConfig,
-			serviceName.orElse( cmdConfig.executionTarget() ) );
-
-		interpreter = new Interpreter(
-			config,
-			params,
-			Optional.of( currInterpreter.logPrefix() ) );
+			interpreter = new Interpreter(
+				config,
+				params,
+				Optional.of( currInterpreter.logPrefix() ) );
+		}
 	}
 
+	/**
+	 * Creates a new instance of ServiceLoader by Jolie code.
+	 *
+	 * @param code the Jolie code of the service to load.
+	 * @param channelDest the channel destination to use to communicate with the service loaded by this
+	 *        loader.
+	 * @param currInterpreter the Jolie interpreter to use to load the service.
+	 * @throws IOException if an IO error occurs while loading the service.
+	 */
 	public JolieServiceLoader( String code, Expression channelDest, Interpreter currInterpreter )
 		throws IOException {
 		super( channelDest );
-		Interpreter.Configuration configuration =
-			Interpreter.Configuration.create( currInterpreter.configuration(), new File( "#native_code_" +
-				SERVICE_LOADER_COUNTER.getAndIncrement() ), new ByteArrayInputStream( code.getBytes() ) );
-		interpreter = new Interpreter( configuration,
-			Optional.empty(), Optional.of( currInterpreter.logPrefix() ) );
+
+		ModuleSource source;
+		try {
+			source = ModuleSource.create(
+				new File( "#native_code_" + SERVICE_LOADER_COUNTER.getAndIncrement() ).toURI(),
+				new ByteArrayInputStream( code.getBytes() ), ServiceNode.DEFAULT_MAIN_SERVICE_NAME );
+
+			Interpreter.Configuration configuration = Interpreter.Configuration.create(
+				currInterpreter.configuration(), source );
+
+			interpreter = new Interpreter( configuration,
+				Optional.empty(), Optional.of( currInterpreter.logPrefix() ) );
+		} catch( FileNotFoundException e ) {
+			throw new IOException(
+				new IllegalStateException( "URI creation should not fail. error: " + e.getMessage() ) );
+		}
+
 	}
 
 	@Override
