@@ -1223,8 +1223,7 @@ public class MetaJolie extends JavaService {
 			configuration.setCheckForMain( false );
 
 			Program program = ParsingUtils.parseProgram(
-				cmdParser.getInterpreterConfiguration().inputStream(),
-				cmdParser.getInterpreterConfiguration().programFilepath().toURI(),
+				cmdParser.getInterpreterConfiguration().source(),
 				cmdParser.getInterpreterConfiguration().charset(),
 				cmdParser.getInterpreterConfiguration().includePaths(),
 				interpreter().configuration().packagePaths(),
@@ -1287,8 +1286,7 @@ public class MetaJolie extends JavaService {
 				new SemanticVerifier.Configuration( cmdParser.getInterpreterConfiguration().executionTarget() );
 			configuration.setCheckForMain( false );
 			Program program = ParsingUtils.parseProgram(
-				cmdParser.getInterpreterConfiguration().inputStream(),
-				cmdParser.getInterpreterConfiguration().programFilepath().toURI(),
+				cmdParser.getInterpreterConfiguration().source(),
 				cmdParser.getInterpreterConfiguration().charset(),
 				cmdParser.getInterpreterConfiguration().includePaths(),
 				interpreter().configuration().packagePaths(),
@@ -1383,8 +1381,7 @@ public class MetaJolie extends JavaService {
 			configuration.setCheckForMain( false );
 
 			Program program = ParsingUtils.parseProgram(
-				cmdParser.getInterpreterConfiguration().inputStream(),
-				cmdParser.getInterpreterConfiguration().programFilepath().toURI(),
+				cmdParser.getInterpreterConfiguration().source(),
 				cmdParser.getInterpreterConfiguration().charset(),
 				cmdParser.getInterpreterConfiguration().includePaths(),
 				interpreter().configuration().packagePaths(),
