@@ -1,12 +1,9 @@
 package jolie.lang.parse.module;
 
 import java.io.IOException;
-import java.io.InputStream;
 import java.net.URI;
 import java.util.Map;
-
 import jolie.lang.parse.ParserException;
-import jolie.lang.parse.Scanner;
 import jolie.lang.parse.ast.Program;
 
 public class Modules {
@@ -49,28 +46,18 @@ public class Modules {
 	 * Note: this method is meant to be used through the execution of jolie program, as it calls static
 	 * crawl method that cache the result
 	 *
-	 * @param configuration
-	 * @param stream jolie code Inputstream
-	 * @param programURI resource URI
-	 * @return
+	 * @param configuration ModuleParsingConfiguration
+	 * @param source jolie module source
+	 * @return parsed result
 	 * @throws ParserException
 	 * @throws IOException
 	 * @throws ModuleException
 	 */
-	public static ModuleParsedResult parseModule( ModuleParsingConfiguration configuration, InputStream stream,
-		URI programURI )
+	public static ModuleParsedResult parseModule( ModuleParsingConfiguration configuration, ModuleSource source )
 		throws ParserException, IOException, ModuleException {
 		ModuleParser parser = new ModuleParser( configuration );
-		ModuleRecord mainRecord = parser.parse(
-			new Scanner( stream, programURI, configuration.charset(), configuration.includeDocumentation() ) );
-		ModuleFinder finder;
-
-		// TODO: This is a hack for Windows. Re-evaluate in the future.
-		if( programURI.toString().contains( "jap:" ) ) {
-			finder = new ModuleFinderDummy();
-		} else {
-			finder = new ModuleFinderImpl( programURI, configuration.packagePaths() );
-		}
+		ModuleRecord mainRecord = parser.parse( source );
+		ModuleFinder finder = new ModuleFinderImpl( configuration.packagePaths() );
 
 		ModuleCrawler.CrawlerResult crawlResult = ModuleCrawler.crawl( mainRecord, configuration, finder );
 
@@ -78,45 +65,6 @@ public class Modules {
 
 		return new ModuleParsedResult( mainRecord.program(), crawlResult.symbolTables() );
 	}
-
-
-
-	/**
-	 * parses jolie's code stream to ModuleParsedResult, which contains executable ast and it's
-	 * symbolTables.
-	 *
-	 * Note: this method is meant to be used to parse jolie code, no cache will be store in memory
-	 *
-	 * @param configuration
-	 * @param stream jolie code Inputstream
-	 * @param programURI resource URI
-	 * @return
-	 * @throws ParserException
-	 * @throws IOException
-	 * @throws ModuleException
-	 */
-	public static ModuleParsedResult parseModuleLocal( ModuleParsingConfiguration configuration, InputStream stream,
-		URI programURI )
-		throws ParserException, IOException, ModuleException {
-		ModuleParser parser = new ModuleParser( configuration );
-		ModuleRecord mainRecord = parser.parse(
-			new Scanner( stream, programURI, configuration.charset(), configuration.includeDocumentation() ) );
-		ModuleFinder finder;
-
-		// TODO: This is a hack for Windows. Re-evaluate in the future.
-		if( programURI.toString().contains( "jap:" ) ) {
-			finder = new ModuleFinderDummy();
-		} else {
-			finder = new ModuleFinderImpl( programURI, configuration.packagePaths() );
-		}
-
-		ModuleCrawler.CrawlerResult crawlResult = ModuleCrawler.crawl( mainRecord, configuration, finder );
-
-		SymbolReferenceResolver.resolve( crawlResult );
-
-		return new ModuleParsedResult( mainRecord.program(), crawlResult.symbolTables() );
-	}
-
 
 	/**
 	 * Clear the module cache entry
